+++ src/data/fitparser.h
@@ -20,7 +20,7 @@ private:
 
 	bool readData(QFile *file, char *data, size_t size);
 	template<class T> bool readValue(CTX &ctx, T &val);
-	bool skipValue(CTX &ctx, size_t size);
+	bool skipValue(CTX &ctx, quint8 size);
 	bool readField(CTX &ctx, Field *field, quint32 &val);
 
 	bool parseHeader(CTX &ctx);
